openedx_unidigital/edxapp_wrapper/backends/teams_q_v1.py
"""
Teams definitions for Open edX Quince release.
"""

from lms.djangoapps.teams.api import (  # pylint: disable=import-error, unused-import
    get_team_by_team_id,
    get_teams_in_teamset,
)
from lms.djangoapps.teams.errors import (  # pylint: disable=import-error, unused-import
    AddToIncompatibleTeamError,
    AlreadyOnTeamInTeamset,
    NotEnrolledInCourseForTeam,
)
from lms.djangoapps.teams.models import (
    CourseTeamMembership,
)  # pylint: disable=import-error, unused-import
