requirements/test.txt
#
# This file is autogenerated by pip-compile with Python 3.8
# by the following command:
#
#    make upgrade
#
asgiref==3.8.1
    # via
    #   -r requirements/base.txt
    #   django
backports-zoneinfo==0.2.1 ; python_version < "3.9"
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
    #   django
    #   djangorestframework
bridgekeeper==0.9
    # via -r requirements/base.txt
click==8.1.7
    # via code-annotations
code-annotations==1.8.0
    # via -r requirements/test.in
coverage[toml]==7.5.1
    # via pytest-cov
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.txt
    #   django-crum
    #   django-filter
    #   djangorestframework
    #   openedx-filters
django-crum==0.7.9
    # via -r requirements/base.txt
django-filter==23.5
    # via
    #   -c requirements/constraints.txt
    #   -r requirements/base.txt
django-rest-framework==0.1.0
    # via -r requirements/base.txt
djangorestframework==3.15.1
    # via
    #   -r requirements/base.txt
    #   django-rest-framework
exceptiongroup==1.2.1
    # via pytest
iniconfig==2.0.0
    # via pytest
jinja2==3.1.4
    # via code-annotations
markupsafe==2.1.5
    # via jinja2
openedx-atlas==0.6.0
    # via -r requirements/base.txt
openedx-filters==1.8.1
    # via -r requirements/base.txt
packaging==24.0
    # via pytest
pbr==6.0.0
    # via stevedore
pluggy==1.5.0
    # via pytest
pytest==8.2.0
    # via
    #   pytest-cov
    #   pytest-django
pytest-cov==5.0.0
    # via -r requirements/test.in
pytest-django==4.8.0
    # via -r requirements/test.in
python-slugify==8.0.4
    # via code-annotations
pyyaml==6.0.1
    # via code-annotations
sqlparse==0.5.0
    # via
    #   -r requirements/base.txt
    #   django
stevedore==5.2.0
    # via code-annotations
text-unidecode==1.3
    # via python-slugify
tomli==2.0.1
    # via
    #   coverage
    #   pytest
typing-extensions==4.11.0
    # via
    #   -r requirements/base.txt
    #   asgiref
web-fragments==2.2.0
    # via -r requirements/base.txt
